memory changes (but still broken)
