+++ memory/lib/cronTask.js
@@ -1,6 +1,8 @@
 const cron = require("node-cron")
 
+let fileStatsCron
+
 module.exports = (cronString, cronTask) => { 
-    const fileStatsCron = cron.schedule(cronString, cronTask)
+    fileStatsCron = cron.schedule(cronString, cronTask)
     fileStatsCron.start()
 }
